+++ covalent_dispatcher/_db/dispatchdb.py
@@ -144,118 +144,18 @@ class DispatchDB:
         else:
             self._dbpath = get_config("user_interface.dispatch_db")
 
-        # dispatch_id is the primary key
+    def _get_data_store(self) -> DataStore:
+        """Return the DataStore instance to write records."""
 
-        # Initialize the db if necessary; sqlite3 raises
-        # sqlite3.OperationalError if table already exists.
-        self.conn = sqlite3.connect(self._dbpath)
-        try:
-            self.conn.execute(
-                "CREATE TABLE dispatches \
-                (dispatch_id text primary key, \
-                result_dict text)"
-            )
-            self.conn.commit()
-        except sqlite3.OperationalError:
-            pass
+        return DataStore(db_URL=f"sqlite+pysqlite:///{self._dbpath}", initialize_db=True)
 
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        if exc_type:
-            self.conn.rollback()
-        else:
-            self.conn.commit()
-        self.conn.close()
-
-        return False
-
-    def _db_dev_path(self):
-        """This is a temporary method for the Result.persist DB path.add()
-
-        TODO - Take this out when result.save and persist have been switched.
-        """
-        sqlite = ".sqlite"
-        return f"sqlite+pysqlite:///{self._dbpath.split(sqlite)[0]}_dev{sqlite}"
-
-    def save_db(self, result_object: Result, **kwargs):
-
-        result_object.save(**kwargs)
+    def save_db(self, result_object: Result):
 
         try:
             # set echo=True only if covalent is started in debug /develop mode `covalent start -d`
             # `initialize_db` flag can be removed as its redundant (sqlalchemy does check if the tables are
             # created or not before inserting/updating data)
-            result_object.persist(DataStore(self._db_dev_path(), initialize_db=True))
-            # result_object.persist()
+            result_object.persist(self._get_data_store())
         except Exception as e:
             app_log.exception(f"Exception occured while saving to DB: {e}.")
-
-    def get(self, dispatch_ids: [] = []) -> List[Tuple[str, str]]:
-        """
-        Retrieve workflows with the given dispatch ids.
-
-        Args:
-            dispatch_ids: A list of dispatch ids for the sought-after workflows.
-
-        Returns:
-            A list of pairs (dispatch_id, [jsonified result dictionary]).
-        """
-        if len(dispatch_ids) > 0:
-            placeholders = "({})".format(", ".join(["?" for i in dispatch_ids]))
-            sql = (
-                "SELECT * FROM dispatches WHERE \
-            dispatch_id in "
-                + placeholders
-            )
-
-            res = self.conn.execute(sql, dispatch_ids).fetchall()
-
-        else:
-            sql = "SELECT * FROM dispatches"
-
-            res = self.conn.execute(sql).fetchall()
-
-        return res
-
-    def upsert(self, dispatch_id: str, result_obj: Result) -> None:
-        """
-        Insert or update the record with the given dispatch_id.
-
-        Args:
-            dispatch_id: The workflow's dispatch_id.
-            result_obj: The Result object for the workflow.
-
-        The Result is turned into a dictionary and stored as json.
-        """
-
-        jsonified_result = encode_result(result_obj)
-
-        try:
-            sql = "INSERT INTO dispatches (dispatch_id, result_dict) VALUES (?, ?)"
-            self.conn.execute(sql, (dispatch_id, jsonified_result))
-            self.conn.commit()
-
-        except sqlite3.IntegrityError:
-            sql = "UPDATE dispatches SET result_dict = ? WHERE dispatch_id = ?"
-            self.conn.execute(sql, (jsonified_result, dispatch_id))
-            self.conn.commit()
-
-        # sql = "INSERT INTO dispatches (dispatch_id, result_dict) VALUES (?, ?) \
-        # ON CONFLICT (dispatch_id) DO UPDATE SET result_dict = excluded.result_dict"
-
-        # self.conn.execute(sql, (dispatch_id, jsonified_result))
-
-    def delete(self, dispatch_ids: []) -> None:
-        """
-        Delete records with the given dispatch ids.
-
-        Args:
-            dispatch_ids: A list of dispatch ids
-        """
-        placeholders = "({})".format(", ".join(["?" for i in dispatch_ids]))
-        sql = "DELETE FROM dispatches WHERE dispatch_id in " + placeholders
-
-        self.conn.execute(sql, dispatch_ids)
-        self.conn.commit()
+            raise